services/tick-ingestion/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0.1
duckdb==1.1.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
PyJWT==2.9.0
slowapi==0.1.9
sentry-sdk[fastapi]==2.14.0
